# platform/api-contracts/inference-gateway.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessagesRequest": {
        "additionalProperties": true,
        "description": "Request body for a native Anthropic ``POST /v1/messages`` call.\n\nMirrors the Anthropic Messages API shape. ``max_tokens`` is required by Anthropic, so\nit is modelled as required here and its absence is a 422 before the governance path\nruns (it also becomes the OpenAI ``max_tokens`` cap that admission enforces). ``system``\naccepts either a plain string or an Anthropic block array. ``extra=\"allow\"`` forwards\nany other Anthropic sampling field (``temperature``, ``top_p``, ``stop_sequences``,\n``tools``, ``tool_choice``, ``stream``, ``metadata``, ...) so nothing is silently dropped.",
        "properties": {
          "data_classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Classification"
          },
          "max_tokens": {
            "title": "Max Tokens",
            "type": "integer"
          },
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Messages",
            "type": "array"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "stop_sequences": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Sequences"
          },
          "stream": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Stream"
          },
          "system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "System"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "tool_choice": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Choice"
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          },
          "top_p": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P"
          }
        },
        "required": [
          "messages",
          "max_tokens"
        ],
        "title": "MessagesRequest",
        "type": "object"
      },
...
      "ResponsesRequest": {
        "additionalProperties": true,
        "description": "Request body for a native OpenAI ``POST /v1/responses`` call.\n\nMirrors the OpenAI Responses API shape. ``input`` is required and accepts either a plain\nstring (translated to a single user message) or an array of input items / messages.\n``instructions`` is prepended as a system message. ``max_output_tokens`` becomes the\nOpenAI ``max_tokens`` cap that admission enforces. Fields with no chat equivalent are\naccepted (``extra=\"allow\"``) but not forwarded to the runtime.\n\nThe stateful fields ``store`` and ``previous_response_id`` are modelled explicitly so the\nhandler can serve them from the response store, or reject them when the store is off,\ninstead of forwarding them into the chat payload where they would be meaningless.",
        "properties": {
          "data_classification": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Classification"
          },
          "input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "Input"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "max_output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Output Tokens"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "previous_response_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Response Id"
          },
          "store": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store"
          },
          "stream": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Stream"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "tool_choice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Choice"
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tools"
          },
          "top_p": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P"
          }
        },
        "required": [
          "input"
        ],
        "title": "ResponsesRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "API key header accepted by gateway middleware when API key authentication is enabled.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "Bearer token accepted by gateway middleware when API key authentication is enabled.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "OpenAI-compatible private inference gateway with sandbox traceability, admission controls, budget enforcement, redacted audit events, and runtime routing to Ollama or vLLM.",
    "title": "Private AI Platform Kit Inference Gateway",
    "version": "0.29.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/messages": {
      "post": {
        "description": "Native Anthropic Messages API endpoint. The request/response are translated to and from the internal OpenAI chat shape and routed through the SAME governance path as POST /v1/chat/completions: model allowlist, admission limits (including the max_tokens cap), prompt secret policy, sandbox budget, output guardrail, and audit. Anthropic requires max_tokens; a request omitting it is rejected. Streaming is supported: stream=true returns the Anthropic event sequence (message_start, content_block_start/delta/stop, message_delta, message_stop) translated from the same governed chat stream, subject to the gateway's streaming admission toggle.",
        "operationId": "createMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessagesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createmessage",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Create a private Anthropic-style message",
        "tags": [
          "inference"
        ]
      }
    },
...
    "/v1/responses": {
      "post": {
        "description": "OpenAI Responses API endpoint. The request/response are translated to and from the internal OpenAI chat shape and routed through the SAME governance path as POST /v1/chat/completions: model allowlist, admission limits (including the max_output_tokens cap), prompt secret policy, sandbox budget, output guardrail, and audit. Server-side state (store=true, previous_response_id) is opt-in via RESPONSES_STORE_ENABLED and requires a sandbox-bound credential when authentication is on; without the store those fields are rejected with stateful_not_supported rather than silently ignored. Streaming is not supported on this endpoint in this release; send stream=false or use POST /v1/chat/completions for OpenAI-shaped streaming.",
        "operationId": "createResponse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponsesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Createresponse",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Create a private OpenAI Responses API response",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/responses/{response_id}": {
      "delete": {
        "operationId": "deleteResponse",
        "parameters": [
          {
            "in": "path",
            "name": "response_id",
            "required": true,
            "schema": {
              "title": "Response Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Deleteresponse",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Delete a stored response",
        "tags": [
          "inference"
        ]
      },
      "get": {
        "operationId": "getResponse",
        "parameters": [
          {
            "in": "path",
            "name": "response_id",
            "required": true,
            "schema": {
              "title": "Response Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getresponse",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Retrieve a stored response (requires the response store)",
        "tags": [
          "inference"
        ]
      }
    },
    "/v1/responses/{response_id}/input_items": {
      "get": {
        "operationId": "listResponseInputItems",
        "parameters": [
          {
            "in": "path",
            "name": "response_id",
            "required": true,
            "schema": {
              "title": "Response Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Listresponseinputitems",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "List the input items of a stored response",
        "tags": [
          "inference"
        ]
      }
    },
...
  },
...
}